Organize Scoreboard Components
