+++ frontend/src/pages/Scoreboard/Round.tsx
@@ -1,6 +1,6 @@
 import { Box } from "@mui/material";
 import { useParams } from "react-router-dom";
-import RoundScoreboard from "../../components/RoundScoreboard";
+import RoundScoreboard from "../../components/Scoreboard/Round";
 
 export default function RoundScoreboardPage() {
   const { round: round } = useParams() as { round: string };
